+++ test/runtests.jl
@@ -6,11 +6,12 @@ using BVDOutbreakSize: REPORT_SCENARIOS,
                        streams_table, comparison_table,
                        nuts_sample, default_adtype,
                        load_observations,
-                       plot_cumulative_cases,
+                       plot_cumulative_cases, plot_density_overlay,
                        plot_prior_predictive,
                        plot_posterior_predictive, plot_pair,
                        plot_start_date_pair, plot_estimate_comparison,
-                       forecast_reported, forecast_table, plot_forecast
+                       forecast_reported, forecast_table, plot_forecast,
+                       forecast_vs_truth, plot_forecast_vs_truth
 using ADTypes: AutoMooncake
 using DataFrames: DataFrame, nrow
 using Distributions: Normal